Temp commit: Add AVAnalyzer

+++ camerax_ios/ios/Classes/common/PointDelegate.swift
@@ -9,14 +9,14 @@ import Foundation
 
 class PointDelegate: PigeonApiDelegatePointApi {
     func pigeonDefaultConstructor(pigeonApi: PigeonApiPointApi, x: Int64, y: Int64) throws -> Point {
-        return Point(x: Int(x), y: Int(y))
+        return Point(x: x, y: y)
     }
     
     func x(pigeonApi: PigeonApiPointApi, pigeonInstance: Point) throws -> Int64 {
-        return Int64(pigeonInstance.x)
+        return pigeonInstance.x
     }
     
     func y(pigeonApi: PigeonApiPointApi, pigeonInstance: Point) throws -> Int64 {
-        return Int64(pigeonInstance.y)
+        return pigeonInstance.y
     }
 }
