+++ camerax_ios/ios/Classes/common/RectDelegate.swift
@@ -9,22 +9,22 @@ import Foundation
 
 class RectDelegate: PigeonApiDelegateRectApi {
     func pigeonDefaultConstructor(pigeonApi: PigeonApiRectApi, left: Int64, top: Int64, right: Int64, bottom: Int64) throws -> Rect {
-        return Rect(left: Int(left), top: Int(top), right: Int(right), bottom: Int(bottom))
+        return Rect(left: left, top: top, right: right, bottom: bottom)
     }
     
     func left(pigeonApi: PigeonApiRectApi, pigeonInstance: Rect) throws -> Int64 {
-        return Int64(pigeonInstance.left)
+        return pigeonInstance.left
     }
     
     func top(pigeonApi: PigeonApiRectApi, pigeonInstance: Rect) throws -> Int64 {
-        return Int64(pigeonInstance.top)
+        return pigeonInstance.top
     }
     
     func right(pigeonApi: PigeonApiRectApi, pigeonInstance: Rect) throws -> Int64 {
-        return Int64(pigeonInstance.right)
+        return pigeonInstance.right
     }
     
     func bottom(pigeonApi: PigeonApiRectApi, pigeonInstance: Rect) throws -> Int64 {
-        return Int64(pigeonInstance.bottom)
+        return pigeonInstance.bottom
     }    
 }
